refactor(runtime): isolate SQL logging from UserContext
