+++ teaql-runtime/src/context.rs
@@ -4,10 +4,7 @@ use std::future::Future;
 
 use std::pin::Pin;
 use std::sync::{Arc, Mutex, OnceLock};
-use std::time::{Duration, SystemTime};
-
-use teaql_core::{EntityDescriptor, Value};
-use teaql_sql::{CompiledQuery, DatabaseKind};
+use std::time::SystemTime;
 
 use crate::EntityRuntimeState;
 use crate::{
@@ -17,12 +14,18 @@ use crate::{
     MetadataStore, ObjectLocation, RawAuditEvent, RawAuditEventSink, RequestPolicy, RuntimeError,
     local_id_generator,
 };
+use teaql_core::{EntityDescriptor, Value};
 
 mod locking;
+mod logging;
 mod pagination;
 mod transaction;
 pub use locking::RemoteLockProvider;
 use locking::next_local_lock_owner;
+pub use logging::{
+    InfoLogEntry, LogPayload, SqlLogEntry, SqlLogOperation, SqlLogOptions, UnifiedLogBuffer,
+    UnifiedLogEntry,
+};
 pub use pagination::{
     ContinuousPageCursor, ContinuousPageCursorStore, IdSetStore, InMemoryContinuousPageCursorStore,
     InMemoryIdSetStore, RetainedIdSet,
@@ -75,118 +78,6 @@ mod active_root_tests {
         assert!(context.require_active_root("Organization").is_err());
         assert!(UserContext::new().require_active_root("Tenant").is_err());
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SqlLogOperation {
-    Select,
-    Insert,
-    Update,
-    Delete,
-    Recover,
-}
-
-impl SqlLogOperation {
-    pub fn is_select(self) -> bool {
-        matches!(self, Self::Select)
-    }
-
-    pub fn is_mutation(self) -> bool {
-        !self.is_select()
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct SqlLogOptions {
-    pub select: bool,
-    pub mutation: bool,
-}
-
-impl Default for SqlLogOptions {
-    fn default() -> Self {
-        Self::all()
-    }
-}
-
-impl SqlLogOptions {
-    pub fn disabled() -> Self {
-        Self {
-            select: false,
-            mutation: false,
-        }
-    }
-
-    pub fn select_only() -> Self {
-        Self {
-            select: true,
-            mutation: false,
-        }
-    }
-
-    pub fn mutation_only() -> Self {
-        Self {
-            select: false,
-            mutation: true,
-        }
-    }
-
-    pub fn all() -> Self {
-        Self {
-            select: true,
-            mutation: true,
-        }
-    }
-
-    pub fn enabled_for(self, operation: SqlLogOperation) -> bool {
-        match operation.is_select() {
-            true => self.select,
-            false => self.mutation,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct SqlLogEntry {
-    pub operation: SqlLogOperation,
-    pub comment: Option<String>,
-    pub purpose: Option<String>,
-    pub audit_reason: Option<String>,
-    pub trace_path: Vec<teaql_core::TraceNode>,
-    pub sql: String,
-    pub params: Vec<Value>,
-    pub debug_sql: String,
-    pub pretty_sql: String,
-    pub started_at: SystemTime,
-    pub ended_at: SystemTime,
-    pub elapsed: Duration,
-    pub result_count: Option<usize>,
-    pub result_type: Option<String>,
-    pub affected_rows: Option<u64>,
-    pub result_summary: String,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct UnifiedLogEntry {
-    pub timestamp: SystemTime,
-    pub user_identifier: Option<String>,
-    pub trace_chain: Vec<teaql_core::TraceNode>,
-    pub payload: LogPayload,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum LogPayload {
-    Sql(SqlLogEntry),
-    Info(InfoLogEntry),
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct InfoLogEntry {
-    pub message: String,
-}
-
-#[derive(Clone, Default)]
-pub struct UnifiedLogBuffer {
-    pub entries: std::sync::Arc<Mutex<Vec<UnifiedLogEntry>>>,
 }
 
 /// Context-owned proof required by the provider SPI. Its private field prevents
@@ -802,196 +693,6 @@ impl UserContext {
         self.i18n_catalog = catalog;
     }
 
-    pub fn with_sql_log_options(mut self, options: SqlLogOptions) -> Self {
-        self.sql_log_options = options;
-        self
-    }
-
-    pub fn set_sql_log_options(&mut self, options: SqlLogOptions) {
-        self.sql_log_options = options;
-    }
-
-    pub fn enable_select_sql_log(&mut self) {
-        self.sql_log_options.select = true;
-    }
-
-    pub fn enable_mutation_sql_log(&mut self) {
-        self.sql_log_options.mutation = true;
-    }
-
-    pub fn disable_select_sql_log(&mut self) {
-        self.sql_log_options.select = false;
-    }
-
-    pub fn disable_mutation_sql_log(&mut self) {
-        self.sql_log_options.mutation = false;
-    }
-
-    pub fn enable_all_sql_log(&mut self) {
-        self.sql_log_options = SqlLogOptions::all();
-    }
-
-    pub fn disable_sql_log(&mut self) {
-        self.sql_log_options = SqlLogOptions::disabled();
-        self.clear_sql_logs();
-    }
-
-    pub fn sql_log_options(&self) -> SqlLogOptions {
-        self.sql_log_options
-    }
-
-    pub fn sql_logs(&self) -> Vec<SqlLogEntry> {
-        self.sql_log_entries
-            .lock()
-            .map(|entries| entries.clone())
-            .unwrap_or_default()
-    }
-
-    pub fn clear_sql_logs(&self) {
-        if let Ok(mut entries) = self.sql_log_entries.lock() {
-            entries.clear();
-        }
-    }
-
-    pub(crate) fn record_sql_log(
-        &self,
-        operation: SqlLogOperation,
-        query: &CompiledQuery,
-        database_kind: DatabaseKind,
-        started_at: SystemTime,
-        ended_at: SystemTime,
-        elapsed: Duration,
-        result_count: Option<usize>,
-        result_type: Option<String>,
-        affected_rows: Option<u64>,
-        trace_chain: Vec<teaql_core::TraceNode>,
-    ) {
-        if !self.sql_log_options.enabled_for(operation) {
-            return;
-        }
-        let debug_sql = query.debug_sql(database_kind);
-        let result_summary = sql_result_summary(
-            operation,
-            result_count,
-            result_type.as_deref(),
-            affected_rows,
-            &debug_sql,
-        );
-
-        let trace_path = canonical_sql_trace_path(
-            operation,
-            &format!("{database_kind:?}").to_ascii_lowercase(),
-            &trace_chain,
-        );
-        let sql_log_entry = SqlLogEntry {
-            operation,
-            comment: trace_value(&trace_chain, teaql_core::TraceKind::Comment),
-            purpose: trace_value(&trace_chain, teaql_core::TraceKind::Purpose),
-            audit_reason: trace_value(&trace_chain, teaql_core::TraceKind::AuditReason),
-            trace_path: trace_path.clone(),
-            sql: query.sql.clone(),
-            params: query.params.clone(),
-            pretty_sql: pretty_sql(&debug_sql),
-            debug_sql: debug_sql.clone(),
-            started_at,
-            ended_at,
-            elapsed,
-            result_summary: result_summary.clone(),
-            result_count,
-            result_type,
-            affected_rows,
-        };
-
-        if let Ok(mut entries) = self.sql_log_entries.lock() {
-            entries.push(sql_log_entry.clone());
-        }
-
-        if let Some(buf) = self.get_resource::<UnifiedLogBuffer>() {
-            if let Ok(mut entries) = buf.entries.lock() {
-                entries.push(UnifiedLogEntry {
-                    timestamp: started_at,
-                    user_identifier: self.user_identifier.clone(),
-                    trace_chain: trace_path.clone(),
-                    payload: LogPayload::Sql(sql_log_entry.clone()),
-                });
-            }
-        }
-
-        crate::log_formatter::LogManager::write_sql_log(&trace_path, &sql_log_entry);
-    }
-
-    pub(crate) fn record_metadata_log(&self, metadata: &teaql_data_service::ExecutionMetadata) {
-        let operation = match metadata.operation {
-            teaql_data_service::DataServiceOperation::Query => SqlLogOperation::Select,
-            teaql_data_service::DataServiceOperation::Insert => SqlLogOperation::Insert,
-            teaql_data_service::DataServiceOperation::Update => SqlLogOperation::Update,
-            teaql_data_service::DataServiceOperation::Delete => SqlLogOperation::Delete,
-            teaql_data_service::DataServiceOperation::Recover => SqlLogOperation::Update,
-            teaql_data_service::DataServiceOperation::Batch => SqlLogOperation::Update,
-            teaql_data_service::DataServiceOperation::Schema => SqlLogOperation::Update,
-        };
-        if !self.sql_log_options.enabled_for(operation) {
-            return;
-        }
-        if let Some(debug_sql) = &metadata.debug_query {
-            let trace_path =
-                canonical_sql_trace_path(operation, &metadata.backend, &metadata.trace_chain);
-            let sql_log_entry = SqlLogEntry {
-                operation,
-                comment: trace_value(&metadata.trace_chain, teaql_core::TraceKind::Comment)
-                    .or_else(|| metadata.comment.clone()),
-                purpose: trace_value(&metadata.trace_chain, teaql_core::TraceKind::Purpose),
-                audit_reason: trace_value(
-                    &metadata.trace_chain,
-                    teaql_core::TraceKind::AuditReason,
-                ),
-                trace_path: trace_path.clone(),
-                sql: metadata.parameterized_query.clone().unwrap_or_default(),
-                params: metadata.params.clone(),
-                pretty_sql: pretty_sql(debug_sql),
-                debug_sql: debug_sql.clone(),
-                started_at: metadata.started_at,
-                ended_at: metadata.ended_at,
-                elapsed: metadata
-                    .ended_at
-                    .duration_since(metadata.started_at)
-                    .unwrap_or_default(),
-                result_count: metadata.result_count,
-                result_type: None, // Not directly available
-                affected_rows: metadata.affected_rows,
-                result_summary: String::new(), // We can synthesize this if needed, or leave it empty/basic
-            };
-
-            // synthesize a summary for the log
-            let mut summary = String::new();
-            if let Some(c) = metadata.result_count {
-                summary = format!("{} rows returned", c);
-            } else if let Some(a) = metadata.affected_rows {
-                summary = format!("{} rows affected", a);
-            }
-
-            let mut final_entry = sql_log_entry;
-            final_entry.result_summary = summary;
-
-            if let Ok(mut entries) = self.sql_log_entries.lock() {
-                entries.push(final_entry.clone());
-            }
-
-            if let Some(buf) = self.get_resource::<UnifiedLogBuffer>() {
-                if let Ok(mut entries) = buf.entries.lock() {
-                    entries.push(UnifiedLogEntry {
-                        timestamp: metadata.started_at,
-                        user_identifier: self.user_identifier.clone(),
-                        trace_chain: trace_path.clone(),
-                        payload: LogPayload::Sql(final_entry.clone()),
-                    });
-                }
-            }
-
-            crate::log_formatter::LogManager::write_sql_log(&trace_path, &final_entry);
-        }
-    }
-
     pub fn language(&self) -> Language {
         self.language
     }
@@ -1302,203 +1003,6 @@ impl UserContext {
             store.remove(key).await;
         }
     }
-}
-
-fn extract_id_from_sql(sql: &str) -> Option<String> {
-    let sql_lower = sql.to_lowercase();
-    let where_idx = sql_lower.find("where")?;
-    let where_clause = &sql_lower[where_idx + 5..];
-
-    let bytes = where_clause.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        if i + 1 < bytes.len() && &bytes[i..i + 2] == b"id" {
-            // Check boundary before
-            let prev_ok = i == 0 || {
-                let prev_char = bytes[i - 1] as char;
-                !prev_char.is_ascii_alphanumeric() && prev_char != '_' && prev_char != '.'
-            };
-            // Check boundary after
-            let next_ok = i + 2 == bytes.len() || {
-                let next_char = bytes[i + 2] as char;
-                !next_char.is_ascii_alphanumeric() && next_char != '_'
-            };
-
-            if prev_ok && next_ok {
-                // Found the standalone "id" word!
-                // Now look for "=" after it
-                let mut j = i + 2;
-                while j < bytes.len() && (bytes[j] as char).is_whitespace() {
-                    j += 1;
-                }
-                if j < bytes.len() && bytes[j] == b'=' {
-                    j += 1;
-                    while j < bytes.len() && (bytes[j] as char).is_whitespace() {
-                        j += 1;
-                    }
-                    // Now extract the value
-                    let mut val_str = String::new();
-                    if j < bytes.len() && bytes[j] == b'\'' {
-                        j += 1; // consume single quote
-                        while j < bytes.len() && bytes[j] != b'\'' {
-                            val_str.push(bytes[j] as char);
-                            j += 1;
-                        }
-                        return Some(val_str);
-                    }
-                    // No else needed — falls through to unquoted parsing
-                    while j < bytes.len() {
-                        let c = bytes[j] as char;
-                        if !c.is_ascii_alphanumeric() && c != '_' && c != '-' {
-                            break;
-                        }
-                        val_str.push(c);
-                        j += 1;
-                    }
-                    if !val_str.is_empty() {
-                        return Some(val_str);
-                    }
-                }
-            }
-        }
-        i += 1;
-    }
-    None
-}
-
-fn sql_result_summary(
-    operation: SqlLogOperation,
-    result_count: Option<usize>,
-    result_type: Option<&str>,
-    affected_rows: Option<u64>,
-    debug_sql: &str,
-) -> String {
-    match operation {
-        SqlLogOperation::Select => {
-            let count = result_count.unwrap_or(0);
-            match count {
-                0 => "MISS".to_owned(),
-                1 => match result_type {
-                    Some(result_type) => extract_id_from_sql(debug_sql)
-                        .map(|id| format!("{result_type}({id})"))
-                        .unwrap_or_else(|| result_type.to_owned()),
-                    None => "row".to_owned(),
-                },
-                _ => match result_type {
-                    Some(result_type) => format!("{count}*{result_type}"),
-                    None => format!("{count}*rows"),
-                },
-            }
-        }
-        _ => {
-            let affected = affected_rows.unwrap_or(0);
-            format!("{affected} UPDATED")
-        }
-    }
-}
-
-fn trace_value(
-    trace_path: &[teaql_core::TraceNode],
-    kind: teaql_core::TraceKind,
-) -> Option<String> {
-    trace_path
-        .iter()
-        .rev()
-        .find(|node| node.kind == kind)
-        .map(|node| node.comment.clone())
-}
-
-fn canonical_sql_trace_path(
-    operation: SqlLogOperation,
-    backend: &str,
-    source: &[teaql_core::TraceNode],
-) -> Vec<teaql_core::TraceNode> {
-    use teaql_core::{TraceKind, TraceNode};
-
-    if source.iter().any(|node| node.kind == TraceKind::Operation)
-        && source.iter().any(|node| node.kind == TraceKind::Provider)
-        && source.iter().any(|node| node.kind == TraceKind::Sql)
-    {
-        return source
-            .iter()
-            .filter(|node| {
-                !matches!(
-                    node.kind,
-                    TraceKind::Comment | TraceKind::Purpose | TraceKind::AuditReason
-                )
-            })
-            .cloned()
-            .collect();
-    }
-
-    let entity = source
-        .iter()
-        .find(|node| !node.entity_type.trim().is_empty())
-        .map(|node| node.entity_type.clone())
-        .unwrap_or_else(|| "unknown".to_owned());
-    let family = if operation.is_select() {
-        "query"
-    } else {
-        "mutation"
-    };
-    let statement = match operation {
-        SqlLogOperation::Select => "select",
-        SqlLogOperation::Insert => "insert",
-        SqlLogOperation::Update => "update",
-        SqlLogOperation::Delete => "delete",
-        SqlLogOperation::Recover => "recover",
-    };
-    let mut path = vec![TraceNode::typed(
-        TraceKind::Operation,
-        entity.clone(),
-        None,
-        family,
-    )];
-    path.push(TraceNode::typed(
-        if operation.is_select() {
-            TraceKind::Request
-        } else {
-            TraceKind::Entity
-        },
-        entity,
-        None,
-        "",
-    ));
-    path.extend(
-        source
-            .iter()
-            .filter(|node| node.kind == TraceKind::Relation)
-            .cloned(),
-    );
-    path.push(TraceNode::typed(
-        TraceKind::Provider,
-        if backend.trim().is_empty() {
-            "unknown"
-        } else {
-            backend
-        },
-        None,
-        "",
-    ));
-    path.push(TraceNode::typed(TraceKind::Sql, statement, None, ""));
-    path
-}
-
-fn pretty_sql(sql: &str) -> String {
-    let mut pretty = sql.to_owned();
-    for keyword in [
-        " FROM ",
-        " WHERE ",
-        " GROUP BY ",
-        " HAVING ",
-        " ORDER BY ",
-        " LIMIT ",
-        " OFFSET ",
-        " RETURNING ",
-    ] {
-        pretty = pretty.replace(keyword, &format!("\n{}", keyword.trim_start()));
-    }
-    pretty.replace(" AND ", "\n  AND ")
 }
 
 #[cfg(test)]
